class - Separando Actions
